Supercharged parsers with NSRegularExpression.
- FMLAZLyricsPageParser no longer relies on UIWebView to parse
lyrics. Lyrics is stripped with regular expressions. Total time
to fetch lyrics is reduced from 10+ seconds to 2 at most.
- FMLLyricsWikiPageParser still partly relies on UIWebView. It is
able to extract the lyrics from the HTML, but the lyrics is stored
in the form of HTML entities, so we need the web view to convert
the entities to proper characters. Fetch time is still dramatically
reduced, down from 10 - 20 seconds to 3 at most.

Also modified FMLController to load lyrics fetching operations
based on the class name stored in user defaults, instead of
using clumsy if blocks. This ease the process of adding new
lyrics sources: Just subclass FMLOperation and add an entry to
Preferences/FetchMyLyrics.plist.

NOTE: I'm trying to replace Preferences/FetchMyLyrics.plist
with elegant hook to Music.app so that user can control
FetchMyLyrics directly in Music.app, which is 1000x more convenient.

P.S. Added attributions.