Use regex to read XML file instead of NSXMLParser.
As a result, the API parser (which rely on NSXMLParser) is no longer
needed, and has been removed.

And to continue the story of the guy I have a crush on...

He once asked me about this very project. I was going to demonstrate
the tweak to him when I realized that I had his picture as my home
screen wallpaper. To top it off, my home button was stuck. Luckily
for me, the button decided to cooperate with me while the whole thing
happened.

Phew. That would have been way awkward.